Backend/requirements.txt
aioredis==2.0.1
asyncpg==0.30.0
fastapi==0.116.1
pydantic_settings==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41
websockets==11.0.3
